Tidy up code and comments.
Hardcode the list of log levels, rather than getting them from the logger
module.  To get them from the logger module, we needed to hardcode a set of
values, anyway, and we'll break elsewhere if the names change (not least, the
configuration format will change).  So I prefer to be explicit about the
hard-coding.