http://code.google.com/p/flaxcode/issues/detail?id=56 http://code.google.com/p/flaxcode/issues/detail?id=71
Added additional checks to see whether the remote process is still
running and to deal with exceptions raised by a filter running in a
remote process.

If the remote process raises an uncaught exception kill it and
restart. This ensures that the next document won't timeout because the
remote process has become unresponsive. It also deal partially with
the problem of growing memory use due (we think) to a problem with the
adobe IFilter.

Added some logging calls.