Rework the way in which flax settings are stored and restored from files. Instead of just pickling (well, shelving) the flax.options object, add a FlaxOptions::to_dict() method, which returns a dict which can then be pickled.  Currently, the DocCollection objects are still put into the dict natively, but everything else placed in the dict is one of (string, list, tuple, dict).  Once the DocCollection objects are resolved, we can start storing the settings as a human-readable format, and can also be more robust to change; we'll be able to read configuration files and upgrade them, since reading them won't rely on the code used to create them being the same as that used to read them.
Update configuration version number to 12.  Remove FlaxOptions.filters
property, which was never used (externally).  Change the
FlaxOptions.formats, FlaxOptions.logger_names and FlaxOptions.languages to
be class properties, since they're dependent on the modules available, not
user configuration.  (Really, they should be available as module
properties, or something like that, but this is a convenient way to
separate them for now, without requiring any of the code using them to be
updated.)  Add FlaxOptions.to_dict() and from_dict(), and convert
make_options() into a FlaxOptions.set_to_defaults() method.

The flax.options object is now created at module import time, and never
overwritten; this removes potential race conditions by stopping it being
accessed before being created.  Instead, the persist module updates the
options in-place.

Add "deleted" attribute to DocCollection objects - this will be used to
mark collections which are in the process of being deleted (but isn't yet
used - I just wanted to get it in now so I don't have to bump the options
file version number again).

CollectionList no longer takes a "formats" object: this was just a
reference to the formats object in flax.options, and as such can more
cleanly be accessed directly (this also makes it obvious that changes in
fax.options.formats will be reflected in any new databases created).